Do not include actorsystem.h via wilson_span.h
